+++ src/app/HISTORY.md
@@ -31,6 +31,9 @@ App Framework Change History
   `create()` method, but before that model has actually been saved and added to
   the list (and before the `add` method has fired). [Ticket #2531400]
 
+* ModelList now allows you to add models to the list even if they were
+  instantiated in another window or another YUI sandbox. [Ticket #2531543]
+
 ### Router (formerly Controller)
 
 * [!] The `Controller` class and `controller` module have been renamed to